添加BOOK_Image_url Book_summary 两个字段在BOOK表里面，同时更新相关的test entity dao

+++ src/test/java/com/turtledove/bookdrift/infrastruct/dao/test/BookDaoTest.java
@@ -18,7 +18,7 @@ import com.turtledove.bookdrift.infrastruct.dao.BookDao;
 public class BookDaoTest extends AbstractTestCase<Book> {
 
 	private static final int VERSION = 2;
-	private static final int PRICE = 39;
+	private static final Double PRICE = (double) 39;
 	private static final String BOOK_PRESS = "人民出版社";
 	private static final String AUTHOR_NAME = "roger";
 	private static final String BOOK_NAME = "程序员的修炼之道";
